+++ LookupAnything/Framework/Lookups/Items/ItemSubject.cs
@@ -673,6 +673,15 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
                     neededFor.Add(I18n.Item_NeededFor_CraftMaster(recipes: string.Join(", ", uncraftedNames)));
             }
 
+            // quests
+            string[] quests = this.GameHelper.GetQuests()
+                .Where(q => q != null && !string.IsNullOrWhiteSpace(q.DisplayText) && q.NeedsItems)
+                .Where(q => q.IsValidItem(obj))
+                .OrderBy(q => q.DisplayText)
+                .Select(q => q.DisplayText).ToArray();
+            if (quests.Any())
+                neededFor.Add(I18n.Item_NeededFor_Quests(quests: string.Join(", ", quests)));
+
             // yield
             if (neededFor.Any())
                 yield return new GenericField(I18n.Item_NeededFor(), string.Join(", ", neededFor));